requirements.txt
anthropic
biopython
streamlit
python-dotenv
